Move configmap, docker_containers, downward_api, expansion and secrets test into common directory.
