+++ test/e2e/common/configmap.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_node
+package common
 
 import (
 	"fmt"
@@ -29,7 +29,6 @@ import (
 )
 
 var _ = framework.KubeDescribe("ConfigMap", func() {
-
 	f := framework.NewDefaultFramework("configmap")
 
 	It("should be consumable from pods in volume [Conformance]", func() {
@@ -80,7 +79,12 @@ var _ = framework.KubeDescribe("ConfigMap", func() {
 		}
 
 		By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
-
+		defer func() {
+			By("Cleaning up the configMap")
+			if err := f.Client.ConfigMaps(f.Namespace.Name).Delete(configMap.Name); err != nil {
+				framework.Failf("unable to delete configMap %v: %v", configMap.Name, err)
+			}
+		}()
 		var err error
 		if configMap, err = f.Client.ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
 			framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
@@ -106,7 +110,7 @@ var _ = framework.KubeDescribe("ConfigMap", func() {
 				Containers: []api.Container{
 					{
 						Name:    containerName,
-						Image:   ImageRegistry[mountTestImage],
+						Image:   "gcr.io/google_containers/mounttest:0.6",
 						Command: []string{"/mt", "--break_on_expected_content=false", "--retry_time=120", "--file_content_in_loop=/etc/configmap-volume/data-1"},
 						VolumeMounts: []api.VolumeMount{
 							{
@@ -121,6 +125,10 @@ var _ = framework.KubeDescribe("ConfigMap", func() {
 			},
 		}
 
+		defer func() {
+			By("Deleting the pod")
+			f.PodClient().Delete(pod.Name, api.NewDeleteOptions(0))
+		}()
 		By("Creating the pod")
 		f.PodClient().CreateSync(pod)
 
@@ -144,7 +152,12 @@ var _ = framework.KubeDescribe("ConfigMap", func() {
 		name := "configmap-test-" + string(uuid.NewUUID())
 		configMap := newConfigMap(f, name)
 		By(fmt.Sprintf("Creating configMap %v/%v", f.Namespace.Name, configMap.Name))
-
+		defer func() {
+			By("Cleaning up the configMap")
+			if err := f.Client.ConfigMaps(f.Namespace.Name).Delete(configMap.Name); err != nil {
+				framework.Failf("unable to delete configMap %v: %v", configMap.Name, err)
+			}
+		}()
 		var err error
 		if configMap, err = f.Client.ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
 			framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
@@ -158,7 +171,7 @@ var _ = framework.KubeDescribe("ConfigMap", func() {
 				Containers: []api.Container{
 					{
 						Name:    "env-test",
-						Image:   ImageRegistry[busyBoxImage],
+						Image:   "gcr.io/google_containers/busybox:1.24",
 						Command: []string{"sh", "-c", "env"},
 						Env: []api.EnvVar{
 							{
@@ -182,6 +195,84 @@ var _ = framework.KubeDescribe("ConfigMap", func() {
 		f.TestContainerOutput("consume configMaps", pod, 0, []string{
 			"CONFIG_DATA_1=value-1",
 		})
+	})
+
+	It("should be consumable in multiple volumes in the same pod", func() {
+		var (
+			name             = "configmap-test-volume-" + string(uuid.NewUUID())
+			volumeName       = "configmap-volume"
+			volumeMountPath  = "/etc/configmap-volume"
+			volumeName2      = "configmap-volume-2"
+			volumeMountPath2 = "/etc/configmap-volume-2"
+			configMap        = newConfigMap(f, name)
+		)
+
+		By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
+		defer func() {
+			By("Cleaning up the configMap")
+			if err := f.Client.ConfigMaps(f.Namespace.Name).Delete(configMap.Name); err != nil {
+				framework.Failf("unable to delete configMap %v: %v", configMap.Name, err)
+			}
+		}()
+		var err error
+		if configMap, err = f.Client.ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
+			framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
+		}
+
+		pod := &api.Pod{
+			ObjectMeta: api.ObjectMeta{
+				Name: "pod-configmaps-" + string(uuid.NewUUID()),
+			},
+			Spec: api.PodSpec{
+				Volumes: []api.Volume{
+					{
+						Name: volumeName,
+						VolumeSource: api.VolumeSource{
+							ConfigMap: &api.ConfigMapVolumeSource{
+								LocalObjectReference: api.LocalObjectReference{
+									Name: name,
+								},
+							},
+						},
+					},
+					{
+						Name: volumeName2,
+						VolumeSource: api.VolumeSource{
+							ConfigMap: &api.ConfigMapVolumeSource{
+								LocalObjectReference: api.LocalObjectReference{
+									Name: name,
+								},
+							},
+						},
+					},
+				},
+				Containers: []api.Container{
+					{
+						Name:  "configmap-volume-test",
+						Image: "gcr.io/google_containers/mounttest:0.6",
+						Args:  []string{"--file_content=/etc/configmap-volume/data-1"},
+						VolumeMounts: []api.VolumeMount{
+							{
+								Name:      volumeName,
+								MountPath: volumeMountPath,
+								ReadOnly:  true,
+							},
+							{
+								Name:      volumeName2,
+								MountPath: volumeMountPath2,
+								ReadOnly:  true,
+							},
+						},
+					},
+				},
+				RestartPolicy: api.RestartPolicyNever,
+			},
+		}
+
+		f.TestContainerOutput("consume configMaps", pod, 0, []string{
+			"content of file \"/etc/configmap-volume/data-1\": value-1",
+		})
+
 	})
 })
 
@@ -208,7 +299,12 @@ func doConfigMapE2EWithoutMappings(f *framework.Framework, uid, fsGroup int64) {
 	)
 
 	By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
-
+	defer func() {
+		By("Cleaning up the configMap")
+		if err := f.Client.ConfigMaps(f.Namespace.Name).Delete(configMap.Name); err != nil {
+			framework.Failf("unable to delete configMap %v: %v", configMap.Name, err)
+		}
+	}()
 	var err error
 	if configMap, err = f.Client.ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
 		framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
@@ -235,7 +331,7 @@ func doConfigMapE2EWithoutMappings(f *framework.Framework, uid, fsGroup int64) {
 			Containers: []api.Container{
 				{
 					Name:  "configmap-volume-test",
-					Image: ImageRegistry[mountTestImage],
+					Image: "gcr.io/google_containers/mounttest:0.6",
 					Args:  []string{"--file_content=/etc/configmap-volume/data-1"},
 					VolumeMounts: []api.VolumeMount{
 						{
@@ -273,7 +369,12 @@ func doConfigMapE2EWithMappings(f *framework.Framework, uid, fsGroup int64) {
 	)
 
 	By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
-
+	defer func() {
+		By("Cleaning up the configMap")
+		if err := f.Client.ConfigMaps(f.Namespace.Name).Delete(configMap.Name); err != nil {
+			framework.Failf("unable to delete configMap %v: %v", configMap.Name, err)
+		}
+	}()
 	var err error
 	if configMap, err = f.Client.ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
 		framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
@@ -306,7 +407,7 @@ func doConfigMapE2EWithMappings(f *framework.Framework, uid, fsGroup int64) {
 			Containers: []api.Container{
 				{
 					Name:  "configmap-volume-test",
-					Image: ImageRegistry[mountTestImage],
+					Image: "gcr.io/google_containers/mounttest:0.6",
 					Args:  []string{"--file_content=/etc/configmap-volume/path/to/data-2"},
 					VolumeMounts: []api.VolumeMount{
 						{
